data_analyzer.py
import pandas as pd
import numpy as np
from sklearn.preprocessing import StandardScaler
from sklearn.cluster import KMeans
import tensorflow as tf
from datetime import datetime
import logging

class DataAnalyzer:
    def __init__(self):
        self.logger = logging.getLogger(__name__)
        self.setup_logging()
        
    def setup_logging(self):
        logging.basicConfig(
            level=logging.INFO,
            format='%(asctime)s [%(levelname)s] %(message)s',
            handlers=[
                logging.FileHandler('data_analysis.log'),
                logging.StreamHandler()
            ]
        )
    
    def preprocess_data(self, data):
        """数据预处理"""
        try:
            # 数据清洗和标准化
            scaler = StandardScaler()
            processed_data = scaler.fit_transform(data)
            return processed_data
        except Exception as e:
            self.logger.error(f"Error preprocessing data: {str(e)}")
            return None
            
    def analyze_patterns(self, data):
        """分析数据模式"""
        try:
            # 使用K-means聚类分析
            kmeans = KMeans(n_clusters=3)
            clusters = kmeans.fit_predict(data)
            return clusters
        except Exception as e:
            self.logger.error(f"Error analyzing patterns: {str(e)}")
            return None
            
    def predict_trends(self, data):
        """预测趋势"""
        try:
            # 使用简单的神经网络进行预测
            model = tf.keras.Sequential([
                tf.keras.layers.Dense(64, activation='relu'),
                tf.keras.layers.Dense(32, activation='relu'),
                tf.keras.layers.Dense(1)
            ])
            # 实现预测逻辑
            return model
        except Exception as e:
            self.logger.error(f"Error predicting trends: {str(e)}")
            return None